VOICE_RECOGNATION/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.3
cryptography==3.4.7
python-dotenv==0.19.0
twilio==7.17.0
SQLAlchemy==1.4.23
